optimize memory usage of simple data implementations

+++ corehunter-base/src/main/java/org/corehunter/data/BiAllelicGenotypeData.java
@@ -19,6 +19,8 @@
 
 package org.corehunter.data;
 
+import org.corehunter.util.CoreHunterConstants;
+
 /**
  * Biallelic data contains the allele score (0/1/2) for markers that have two and only two alleles.
  * If any of the markers have three or more alleles, such as classic SSR data, then the general
@@ -33,13 +35,15 @@ public interface BiAllelicGenotypeData extends FrequencyGenotypeData {
      * Get the allele score of the marker for the given entry. The marker scores are encoded as 0, 1 or 2.
      * 0 is used for homozygote for one allele and 2 is used for homozygote for the other allele. 1 is used
      * to denote heterozygotes. It is not important which allele is encoded as 0 or 2, as long as it is consistent
-     * for a specific marker. If the allele score is missing this method returns <code>null</code>.
+     * for a specific marker. If the allele score is missing this method returns
+     * {@link CoreHunterConstants#MISSING_ALLELE_SCORE}.
      *
      * @param id    the id of the entry, must be one of the IDs returned by {@link #getIDs()}
      * @param markerIndex the index of the marker within the range 0 to n-1, where n is the total number of markers and
      *                    is returned by {@link #getNumberOfMarkers()}
-     * @return the allele score of the marker for the given entry (0, 1 or 2); <code>null</code> if missing
+     * @return the allele score of the marker for the given entry (0, 1 or 2);
+     *         {@link CoreHunterConstants#MISSING_ALLELE_SCORE} if missing
      */
-    public Integer getAlleleScore(int id, int markerIndex);
+    public byte getAlleleScore(int id, int markerIndex);
     
 }
